requirements.txt
asgiref==3.4.1
astroid==2.3.3
beautifulsoup4==4.10.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.9
Django==3.2.10
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
djangorestframework==3.12.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.4.0
pytz==2021.3
requests==2.22.0
six==1.13.0
soupsieve==2.3.1
sqlparse==0.4.2
sslcommerz-lib==1.0
sslcommerz-python==0.0.7
typed-ast==1.4.0
typing-extensions==4.0.1
urllib3==1.25.7
wrapt==1.11.2
